smart-library-ai/recommendation/requirements.txt
# 推荐系统依赖

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# 数据库
SQLAlchemy>=2.0.0
PyMySQL>=1.1.0

# 环境变量
python-dotenv>=1.0.0

# Redis 缓存（可选）
redis>=5.0.0

# MySQL 认证支持
cryptography>=46.0.0
